Refactoring moved files
